Se hace cambio al modelo completo de arquitectura clean funcional

+++ ProyectoCs/ProyectoCs/src/main/java/com/example/ProyectoCs/domain/repository/RecomendacionRepository.java
@@ -5,6 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface RecomendacionRepository extends JpaRepository<Recomendacion, Integer> {
-    // Puedes agregar métodos personalizados de consulta si es necesario
+public interface RecomendacionRepository extends JpaRepository<Recomendacion, Long> {
 }
