+++ ProyectoCs/ProyectoCs/src/main/java/com/example/ProyectoCs/domain/repository/TipoAlojamientoRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface TipoAlojamientoRepository extends JpaRepository<TipoAlojamiento, Integer> {
+public interface TipoAlojamientoRepository extends JpaRepository<TipoAlojamiento, Long> {
 }
